backend/myLubd/src/requirements.txt
asgiref==3.8.1
cachetools==5.5.2
certifi==2024.8.30
charset-normalizer==3.4.0
defusedxml==0.7.1
Django==4.2.16
django-cors-headers==4.4.0
django-dbbackup==4.2.1
django-debug-toolbar==4.4.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
gunicorn==21.2.0
httpie==3.2.4
httplib2==0.22.0
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
oauthlib==3.2.2
packaging==24.2
pillow==10.4.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.2.3
PySocks==1.7.1
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
sqlparse==0.5.2
typing_extensions==4.12.2
urllib3==2.2.3
django-dbbackup==4.2.1
django-filter==25.1